# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News-aggregator with User Authentication(Sanctum) API Documentation",
        "description": "Swagger OpenAPI Description",
        "contact": {
            "email": "farooqaziz64@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/news/list": {
            "get": {
                "operationId": "695f0c701b824a1a25a152040faac400",
                "parameters": [
                    {
                        "name": "source",
                        "in": "query",
                        "description": "source",
                        "example": "bbc-news"
                    },
                    {
                        "name": "published_at[from]",
                        "in": "query",
                        "description": "published_at from",
                        "example": "2023-12-03"
                    },
                    {
                        "name": "published_at[to]",
                        "in": "query",
                        "description": "published_at to",
                        "example": "2023-12-05"
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "category",
                        "example": "sport"
                    },
                    {
                        "name": "preference",
                        "in": "query",
                        "description": "preferences",
                        "example": "my_favorite"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                }
            }
        },
        "/api/v1/news/search": {
            "get": {
                "operationId": "3671e0baa075f8fd5bb146a7913c2ea2",
                "parameters": [
                    {
                        "name": "text",
                        "in": "query",
                        "description": "text",
                        "example": "movie"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                }
            }
        },
...
    }
}